% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\name{Calculae_Cossim_catalogue_RepIndel}
\alias{Calculae_Cossim_catalogue_RepIndel}
\title{Compute cosine similarity between one sample indel profile and MMR gene knockouts}
\usage{
Calculae_Cossim_catalogue_RepIndel(singlesample_indel_catalogue, Sig)
}
\arguments{
\item{singlesample_indel_catalogue}{indel catalogue of one sample}

\item{Sig}{mismatch repair gene knockout indel signatures}
}
\value{
cosine similarity between given sub profiles and MMR gene knockout sub profiles
}
\description{
Compute cosine similarity between one sample indel profile and MMR gene knockouts
}
