% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\name{cos_similarity}
\alias{cos_similarity}
\title{Compute cosine similarity between two vectors}
\usage{
cos_similarity(v1, v2)
}
\arguments{
\item{v1}{vector 1}

\item{v2}{vector 2}
}
\value{
cosine similarity
}
\description{
Compute cosine similarity between two vectors
}
