% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_catalogues.R
\name{findsmallestrep}
\alias{findsmallestrep}
\title{This finds the smallest repeating subunit of the deletion}
\usage{
findsmallestrep(d)
}
\description{
This finds the smallest repeating subunit of the deletion
}
