% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_catalogues.R
\name{gen_indelmuttype_MMRD}
\alias{gen_indelmuttype_MMRD}
\title{Generate 45 channel indel catalogue}
\usage{
gen_indelmuttype_MMRD(muts_list, Sample_col, muttype_col)
}
\arguments{
\item{muts_list}{A list of indels with flanking sequence information: Chrom, Pos, Ref, Alt, Sample column}

\item{Sample_col}{Sample column name}

\item{muttype_col}{indel classification column name}
}
\value{
Indel catalogue
}
\description{
Generate 45 channel indel catalogue
}
