% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_catalogues.R
\name{GenCatalogue}
\alias{GenCatalogue}
\title{Generate 96 channel catalogue for substitutions}
\usage{
GenCatalogue(CTsubs, SampleCol)
}
\arguments{
\item{CTsubs}{A list of substitutions: Chrom, Pos, Ref, Alt, Sample column}

\item{SampleCol}{Sample column}
}
\value{
96 channel catalogue for substitutions
}
\description{
Generate 96 channel catalogue for substitutions
}
