% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\name{Generate_CossimVector_SingleSample_RepIndel}
\alias{Generate_CossimVector_SingleSample_RepIndel}
\title{Compute cosine similarity between one sample and MMR gene knockouts}
\usage{
Generate_CossimVector_SingleSample_RepIndel(
  SingleSample_classified_indels,
  Sig = MMRKO.indelsig
)
}
\arguments{
\item{SingleSample_classified_indels}{classified indel list}

\item{Sig}{mismatch repair gene knockout indel signatures}
}
\value{
cosine similarity between given sub profiles and MMR gene knockout sub profiles
}
\description{
Compute cosine similarity between one sample and MMR gene knockouts
}
