% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_catalogues.R
\name{indel_classifier}
\alias{indel_classifier}
\title{Generate indel catalogue for substitutions}
\usage{
indel_classifier(indels)
}
\arguments{
\item{CTsubs}{A list of substitutions: Chrom, Pos, Ref, Alt, Sample column}

\item{SampleCol}{Sample column}
}
\value{
96 channel catalogue for substitutions
}
\description{
Generate indel catalogue for substitutions
}
