% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_catalogues.R
\name{mh_indel_v2}
\alias{mh_indel_v2}
\title{find mh indels}
\usage{
mh_indel_v2(indel.df)
}
\arguments{
\item{indel.df}{A list of indels with flanking sequence information: Chrom, Pos, Ref, Alt, Sample column}
}
\value{
annotated indel list
}
\description{
find mh indels
}
