% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_catalogues.R
\name{mhcaller}
\alias{mhcaller}
\title{micro-homology between the deleted sequence and 3' sequence (number of bases)}
\usage{
mhcaller(d, prime3)
}
\arguments{
\item{d}{A list of indels with flanking sequence information: Chrom, Pos, Ref, Alt, Sample column}
}
\value{
annotated indels
}
\description{
micro-homology between the deleted sequence and 3' sequence (number of bases)
}
