% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.classify.R
\docType{data}
\name{MMRDclassifier}
\alias{MMRDclassifier}
\title{Default MMRDclassifier.}
\format{
R object list of 30
}
\usage{
MMRDclassifier
}
\description{
MMRDclassifier trained on ~270 colorectal cancers.
}
\keyword{datasets}
