% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.classify.R
\name{MMRDetect.classify}
\alias{MMRDetect.classify}
\title{MMRDetect classifier}
\usage{
MMRDetect.classify(mutationVariable, classifier = MMRDclassifier)
}
\arguments{
\item{mutationVariable}{A list of input variables,"Del_rep_mean","RepIndel_num","MMR_sum","maxcossim"}

\item{classifier}{provided classifier}
}
\value{
classification result
}
\description{
MMRDetect classifier
}
