% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\name{MMRDetect.compute.Repindel.similarity}
\alias{MMRDetect.compute.Repindel.similarity}
\title{Compute cosine similarity between given indel profile and MMR gene knockout indel profilesct}
\usage{
MMRDetect.compute.Repindel.similarity(
  indels,
  tissue_type,
  MMR_sig_indel = MMRKO_indelsig
)
}
\arguments{
\item{indels}{list of indels}

\item{tissue_type}{tissue type}

\item{MMR_sig_indel}{mismatch repair gene knockout indel signatures}
}
\value{
cosine similarity between given indel list and MMR gene knockout indel profiles
}
\description{
Compute cosine similarity between given indel profile and MMR gene knockout indel profilesct
}
