% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\name{MMRDetect.compute.Repindelcatalogue.similarity}
\alias{MMRDetect.compute.Repindelcatalogue.similarity}
\title{Compute cosine similarity between given indel profile and MMR gene knockout indel profiles}
\usage{
MMRDetect.compute.Repindelcatalogue.similarity(
  indel_cat,
  tissue_type,
  MMR_sig_indel = MMRKO_indelsig
)
}
\arguments{
\item{indel_cat}{indels catalogue}

\item{tissue_type}{tissue type}

\item{MMR_sig_indel}{mismatch repair gene knockout indel signatures}
}
\value{
cosine similarity between given indel profile and MMR gene knockout indel profiles
}
\description{
Compute cosine similarity between given indel profile and MMR gene knockout indel profiles
}
