% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\name{MMRDetect.compute.subcatalogue.exposure}
\alias{MMRDetect.compute.subcatalogue.exposure}
\title{Compute substitution signature exposure for a given catalogue}
\usage{
MMRDetect.compute.subcatalogue.exposure(
  sub_cat,
  tissue_type,
  MMR_subsig96 = MMRKO_subsig,
  tissue_subsig96
)
}
\arguments{
\item{sub_cat}{substitution catalogue}

\item{tissue_type}{tissue type}

\item{MMR_subsig96}{mismatch repair gene knockout substitution signatures}

\item{tissue_subsig96}{tissue-specific substitution signatures}
}
\value{
Substitution signature exposure
}
\description{
Compute substitution signature exposure for a given catalogue
}
