% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\name{MMRDetect.compute.subcatalogue.similarity}
\alias{MMRDetect.compute.subcatalogue.similarity}
\title{Compute cosine similarity between given sub profiles and MMR gene knockout sub profiles}
\usage{
MMRDetect.compute.subcatalogue.similarity(mcat, scat = MMRKO_subsig)
}
\arguments{
\item{mcat}{substitution catalogue}

\item{scat}{mismatch repair gene knockout sub signatures}
}
\value{
cosine similarity between given sub profiles and MMR gene knockout sub profiles
}
\description{
Compute cosine similarity between given sub profiles and MMR gene knockout sub profiles
}
