% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.train.R
\name{MMRDetect.train}
\alias{MMRDetect.train}
\title{Train MMRDetect}
\usage{
MMRDetect.train(mutationVariable, classification)
}
\arguments{
\item{mutationVariable}{A list of input variables,"Del_rep_mean","RepIndel_num","MMR_sum","maxcossim"}

\item{classification}{Sample classification}
}
\value{
trained model
}
\description{
Train MMRDetect
}
