% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\docType{data}
\name{MMRKO_indelsig}
\alias{MMRKO_indelsig}
\title{Indel signature of MMR gene knockouts}
\format{
A data frame with nine variables:
\describe{
\item{\code{indelsubtype}}{Indel types}
\item{\code{MLH1}}{MLH1 KO}
\item{\code{MSH2}}{MSH2 KO}
\item{\code{MSH6}}{MSH6 KO}
\item{\code{PMS2}}{PM2 KO}
}
}
\usage{
MMRKO_indelsig
}
\description{
Indel signature of MMR gene knockouts
}
\keyword{datasets}
