% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRDetect.compute.variables.R
\name{SigCossim}
\alias{SigCossim}
\title{Compute cosine similarity between a set of catalogue and a given signatures}
\usage{
SigCossim(mcat, sigx)
}
\arguments{
\item{mcat}{mutation catalogue}

\item{sigx}{a given signature}
}
\value{
a list of cosine similarity
}
\description{
Compute cosine similarity between a set of catalogue and a given signatures
}
