% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_catalogues.R
\name{subtype_indel_mmrd}
\alias{subtype_indel_mmrd}
\title{classify indels}
\usage{
subtype_indel_mmrd(indel.df)
}
\arguments{
\item{indel.df}{A list of indels with flanking sequence information: Chrom, Pos, Ref, Alt, Sample column}
}
\value{
96 channel catalogue for substitutions
}
\description{
classify indels
}
